moved subdomain routing logic to module
