+++ src/namespace.rs
@@ -1,110 +1,22 @@
-use std::{collections::BTreeMap, collections::HashMap, sync::Arc};
+use std::{collections::BTreeMap, sync::Arc};
 
-use crate::{router::Selection, subdomain::Subdomain, Middleware};
-
-enum Param {
-    Param(String),
-    String(String),
-}
-
-struct Holder<T> {
-    data: T,
-    map: Vec<Param>,
-}
-
-struct Match<'a, T> {
-    data: &'a T,
-    params: BTreeMap<&'a String, String>,
-}
-
-impl<T> Holder<T> {
-    pub fn new(domain: &str, data: T) -> Holder<T> {
-        let map = domain
-            .split('.')
-            .rev()
-            .map(|p| {
-                if p.starts_with(":") {
-                    Param::Param(p[1..].to_owned())
-                } else {
-                    Param::String(p.to_owned())
-                }
-            })
-            .collect();
-        Holder { data, map }
-    }
-
-    pub fn compare(&self, parts: &Vec<&str>) -> Option<Match<'_, T>> {
-        if self.map.len() != parts.len() {
-            return None;
-        }
-        let mut m = Match {
-            data: &self.data,
-            params: BTreeMap::new(),
-        };
-        for (url_part, subdomain_part) in parts.iter().zip(&self.map) {
-            match subdomain_part {
-                Param::Param(param_name) => {
-                    m.params.insert(param_name, url_part.to_string());
-                }
-                Param::String(exact_name) => {
-                    if exact_name == "*" {
-                        continue;
-                    } else if url_part != exact_name {
-                        return None;
-                    }
-                }
-            }
-        }
-        return Some(m);
-    }
-
-    pub fn data(&mut self) -> &mut T {
-        &mut self.data
-    }
-}
+use crate::{
+    router::Selection, subdomain::Subdomain, subdomain_router::router::SubdomainRouter, Middleware,
+};
 
 pub struct Namespace<State> {
-    subdomain_map: SubdomainMapper<Subdomain<State>>,
-}
-
-struct SubdomainMapper<T> {
-    pub subdomains: HashMap<String, Holder<T>>,
-}
-
-impl<T> SubdomainMapper<T> {
-    pub fn new() -> Self {
-        Self {
-            subdomains: HashMap::new(),
-        }
-    }
-
-    pub fn add(&mut self, subdomain: &str, element: T) -> &mut T {
-        self.subdomains
-            .entry(subdomain.to_owned())
-            .or_insert_with(|| Holder::new(subdomain, element))
-            .data()
-    }
-
-    pub fn recognize(&self, domain: &str) -> Option<Match<'_, T>> {
-        let domain = domain.split('.').rev().collect::<Vec<&str>>();
-        for value in self.subdomains.values() {
-            if let Some(subdomain) = value.compare(&domain) {
-                return Some(subdomain);
-            }
-        }
-        None
-    }
+    router: SubdomainRouter<Subdomain<State>>,
 }
 
 impl<State: Clone + Send + Sync + 'static> Namespace<State> {
     pub fn new() -> Self {
         Self {
-            subdomain_map: SubdomainMapper::new(),
+            router: SubdomainRouter::new(),
         }
     }
 
     pub fn add(&mut self, subdomain: String, router: Subdomain<State>) -> &mut Subdomain<State> {
-        self.subdomain_map.add(&subdomain, router)
+        self.router.add(&subdomain, router)
     }
 
     pub fn route(
@@ -125,7 +37,7 @@ impl<State: Clone + Send + Sync + 'static> Namespace<State> {
                 .to_owned()
         };
 
-        match self.subdomain_map.recognize(&domain) {
+        match self.router.recognize(&domain) {
             Some(data) => {
                 let subdomain = data.data;
                 let params = data.params;
